refactor ConsentCheckbox into separate component
